Migrate much of raven to seperate lib file, for gtk version
